refactor(metadata-core,objectql): sink both write-verb dispatch predicates into metadata-core (#5619)
Move `engine-delete-dispatch.ts` (#4550) and `engine-update-dispatch.ts` (#5480)
from `packages/objectql/src/` to `packages/metadata-core/src/` unchanged, and
keep the original objectql paths as re-export shims so no caller, no public
export and no pinned call site moves.

`@objectstack/objectql` depends on `@objectstack/metadata-protocol`, so that
package's 13 fake engines could not import either predicate without closing a
dependency cycle turbo 2.10.7 refuses outright. `@objectstack/metadata-core` is
a package both sides already depend on and which depends on neither, so the
sink is the one route that pins those doubles without inventing an edge.

- 13 metadata-protocol test files: both fake write verbs now open with the
  producer's own predicate, imported from `@objectstack/metadata-core`.
- 26 (file, verb) ledger entries deleted from
  `scripts/engine-double-contract.baseline.json`; the gate's counts move from
  37 pinned / 165 ledger / 2 exempt to 63 pinned / 139 ledger / 2 exempt.
- The gate's two slices now accept `@objectstack/metadata-core` alongside
  `@objectstack/objectql`, and its remediation message names the right one.
- Six ledger entries whose `closes` prescribed exactly this sink ("tracked as
  #5619") are rewritten: the blocker is gone, what remains is the one-line pin,
  tracked as #5855.
