+++ packages/metadata-protocol/src/protocol.code-only-types.test.ts
@@ -29,6 +29,14 @@
  * cannot be tested against a harness that mocks `saveMetaItem`.
  */
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
+// [#5619] The producer's OWN write-verb dispatch decisions (#4550 delete /
+// #5480 update), so the fake engine below cannot accept a call ObjectQL
+// refuses. Imported from `@objectstack/metadata-core` and not from
+// `@objectstack/objectql`: objectql DEPENDS ON this package, so that import
+// would close a dependency cycle turbo rejects outright — which is why all 26
+// of this package's (file, verb) pairs sat in the gate's DEBT ledger until
+// #5619 sank the two predicates into a package both sides already depend on.
+import { assertEngineDeleteDispatch, assertEngineUpdateDispatch } from '@objectstack/metadata-core';
 import { DEFAULT_METADATA_TYPE_REGISTRY } from '@objectstack/spec/kernel';
 import { ObjectStackProtocolImplementation } from './protocol.js';
 import { resetEnvWritableMetadataTypes } from './sys-metadata-repository.js';
@@ -108,8 +116,16 @@ function makeStubEngine(artifacts: Array<{ type: string; name: string }> = []) {
             rows.set(keyOf(data), row);
             return { id: row.id };
         },
-        async update(_t: string) { writes.push({ op: 'update', table: _t }); return { id: null }; },
-        async delete(_t: string) { writes.push({ op: 'delete', table: _t }); return { deleted: 0 }; },
+        async update(_t: string, data: Record<string, unknown>, opts?: Record<string, unknown>) {
+            assertEngineUpdateDispatch(data, opts);
+            writes.push({ op: 'update', table: _t });
+            return { id: null };
+        },
+        async delete(_t: string, opts?: Record<string, unknown>) {
+            assertEngineDeleteDispatch(opts);
+            writes.push({ op: 'delete', table: _t });
+            return { deleted: 0 };
+        },
         registry: {
             registerItem: () => {},
             registerObject: () => {},
