+++ packages/metadata-protocol/src/protocol.read-decorations.test.ts
@@ -22,6 +22,14 @@
  * future third decoration key fails here instead of in a production boot log.
  */
 import { describe, expect, it } from 'vitest';
+// [#5619] The producer's OWN write-verb dispatch decisions (#4550 delete /
+// #5480 update), so the fake engine below cannot accept a call ObjectQL
+// refuses. Imported from `@objectstack/metadata-core` and not from
+// `@objectstack/objectql`: objectql DEPENDS ON this package, so that import
+// would close a dependency cycle turbo rejects outright — which is why all 26
+// of this package's (file, verb) pairs sat in the gate's DEBT ledger until
+// #5619 sank the two predicates into a package both sides already depend on.
+import { assertEngineDeleteDispatch, assertEngineUpdateDispatch } from '@objectstack/metadata-core';
 import { FlowSchema } from '@objectstack/spec/automation';
 import { METADATA_READ_DECORATIONS } from '@objectstack/spec/kernel';
 import { ObjectStackProtocolImplementation, stripReadDecorations } from './index.js';
@@ -71,6 +79,7 @@ function makeStubEngine() {
             return { id: row.id };
         },
         async update(table: string, data: Record<string, unknown>, opts: { where: Record<string, unknown> }) {
+            assertEngineUpdateDispatch(data, opts);
             if (table !== 'sys_metadata') return { id: null };
             const found = findRow(opts.where);
             if (!found) return { id: null };
@@ -79,7 +88,10 @@ function makeStubEngine() {
             rows.set(keyOf(merged), merged);
             return { id: found.row.id };
         },
-        async delete() { return { deleted: 0 }; },
+        async delete(_t: string, opts?: Record<string, unknown>) {
+            assertEngineDeleteDispatch(opts);
+            return { deleted: 0 };
+        },
         async transaction<T>(cb: (ctx: any) => Promise<T>): Promise<T> { return cb(undefined); },
         async syncObjectSchema() { /* no DDL in this stub */ },
         registry: {
