+++ packages/metadata-protocol/src/protocol.save-union-issues.test.ts
@@ -23,6 +23,14 @@
  * cannot use a harness that mocks `saveMetaItem`.
  */
 import { describe, expect, it } from 'vitest';
+// [#5619] The producer's OWN write-verb dispatch decisions (#4550 delete /
+// #5480 update), so the fake engine below cannot accept a call ObjectQL
+// refuses. Imported from `@objectstack/metadata-core` and not from
+// `@objectstack/objectql`: objectql DEPENDS ON this package, so that import
+// would close a dependency cycle turbo rejects outright — which is why all 26
+// of this package's (file, verb) pairs sat in the gate's DEBT ledger until
+// #5619 sank the two predicates into a package both sides already depend on.
+import { assertEngineDeleteDispatch, assertEngineUpdateDispatch } from '@objectstack/metadata-core';
 import { getMetadataTypeSchema } from '@objectstack/spec/kernel';
 import { ObjectStackProtocolImplementation, zodIssuesToMetadataIssues } from './protocol.js';
 
@@ -52,8 +60,14 @@ function makeProtocol() {
             rows.set(keyOf(data), row);
             return { id: row.id };
         },
-        async update() { return { id: null }; },
-        async delete() { return { deleted: 0 }; },
+        async update(_t: string, data: Record<string, unknown>, opts?: Record<string, unknown>) {
+            assertEngineUpdateDispatch(data, opts);
+            return { id: null };
+        },
+        async delete(_t: string, opts?: Record<string, unknown>) {
+            assertEngineDeleteDispatch(opts);
+            return { deleted: 0 };
+        },
         registry: { registerItem: () => {}, registerObject: () => {} },
     };
     const protocol: any = new ObjectStackProtocolImplementation(engine, () => new Map());
