+++ packages/metadata-protocol/src/protocol.stored-conversions.test.ts
@@ -16,6 +16,14 @@
  * years ago under an older protocol.
  */
 import { describe, expect, it } from 'vitest';
+// [#5619] The producer's OWN write-verb dispatch decisions (#4550 delete /
+// #5480 update), so the fake engine below cannot accept a call ObjectQL
+// refuses. Imported from `@objectstack/metadata-core` and not from
+// `@objectstack/objectql`: objectql DEPENDS ON this package, so that import
+// would close a dependency cycle turbo rejects outright — which is why all 26
+// of this package's (file, verb) pairs sat in the gate's DEBT ledger until
+// #5619 sank the two predicates into a package both sides already depend on.
+import { assertEngineDeleteDispatch, assertEngineUpdateDispatch } from '@objectstack/metadata-core';
 import { ObjectStackProtocolImplementation } from './protocol.js';
 
 interface Row {
@@ -56,8 +64,14 @@ function makeStubEngine(seedRows: Array<Partial<Row> & { type: string; name: str
             return rows.filter((r) => matches(r, opts.where));
         },
         async insert() { return { id: 'x' }; },
-        async update() { return { id: 'x' }; },
-        async delete() { return { deleted: 0 }; },
+        async update(_t: string, data: Record<string, unknown>, opts?: Record<string, unknown>) {
+            assertEngineUpdateDispatch(data, opts);
+            return { id: 'x' };
+        },
+        async delete(_t: string, opts?: Record<string, unknown>) {
+            assertEngineDeleteDispatch(opts);
+            return { deleted: 0 };
+        },
         registry: {
             listItems: () => [],
             isPackageDisabled: () => false,
