+++ packages/metadata-protocol/src/sys-metadata-repository.draft-drain.test.ts
@@ -27,6 +27,14 @@
  */
 
 import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
+// [#5619] The producer's OWN write-verb dispatch decisions (#4550 delete /
+// #5480 update), so the fake engine below cannot accept a call ObjectQL
+// refuses. Imported from `@objectstack/metadata-core` and not from
+// `@objectstack/objectql`: objectql DEPENDS ON this package, so that import
+// would close a dependency cycle turbo rejects outright — which is why all 26
+// of this package's (file, verb) pairs sat in the gate's DEBT ledger until
+// #5619 sank the two predicates into a package both sides already depend on.
+import { assertEngineDeleteDispatch, assertEngineUpdateDispatch } from '@objectstack/metadata-core';
 import { readFileSync } from 'node:fs';
 import { SysMetadataRepository } from './sys-metadata-repository.js';
 
@@ -158,12 +166,14 @@ function makeFakeEngine() {
       data: Record<string, unknown>,
       opts: { where: Record<string, unknown> },
     ) {
+      assertEngineUpdateDispatch(data, opts);
       const found = findRow(opts.where);
       if (!found) throw new Error('not found');
       rows.set(found.key, { ...found.row, ...data });
       return { id: found.row.id as string };
     },
     async delete(_t: string, opts: { where: Record<string, unknown> }) {
+      assertEngineDeleteDispatch(opts);
       const found = findRow(opts.where);
       if (draftDeleteFailure && found?.row.state === 'draft') throw draftDeleteFailure();
       if (!found) return { deleted: 0 };
