+++ packages/metadata-protocol/src/sys-metadata-repository.recorded-by.test.ts
@@ -17,6 +17,14 @@
  */
 
 import { describe, it, expect, beforeEach } from 'vitest';
+// [#5619] The producer's OWN write-verb dispatch decisions (#4550 delete /
+// #5480 update), so the fake engine below cannot accept a call ObjectQL
+// refuses. Imported from `@objectstack/metadata-core` and not from
+// `@objectstack/objectql`: objectql DEPENDS ON this package, so that import
+// would close a dependency cycle turbo rejects outright — which is why all 26
+// of this package's (file, verb) pairs sat in the gate's DEBT ledger until
+// #5619 sank the two predicates into a package both sides already depend on.
+import { assertEngineDeleteDispatch, assertEngineUpdateDispatch } from '@objectstack/metadata-core';
 import { SysMetadataRepository } from './sys-metadata-repository.js';
 
 interface Row { [k: string]: unknown }
@@ -75,12 +83,14 @@ function makeFakeEngine() {
       return { id: row.id as string };
     },
     async update(_t: string, data: Record<string, unknown>, opts: { where: Record<string, unknown> }) {
+      assertEngineUpdateDispatch(data, opts);
       const found = findRow(opts.where);
       if (!found) throw new Error('not found');
       rows.set(found.key, { ...found.row, ...data });
       return { id: found.row.id as string };
     },
     async delete(_t: string, opts: { where: Record<string, unknown> }) {
+      assertEngineDeleteDispatch(opts);
       const found = findRow(opts.where);
       if (!found) return { deleted: 0 };
       rows.delete(found.key);
